Wrote code to display the message 'Hello World'
The ccode currently displays 'Hello World' on the left side of the screen.
The message is moved down one line every minute and wraps around when it
reaches the bottom of the screen. The next step will be to figure out how
the clock information variables work so that I can change the message to
the appropriate information.